#25 Refactoring multiple screens for better readability - WIP
