+++ app/src/main/java/de/fampopprol/dhbwhorb/ui/screen/GradesScreen.kt
@@ -6,62 +6,43 @@
 
 package de.fampopprol.dhbwhorb.ui.screen
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Star
 import androidx.compose.material3.Button
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.ExposedDropdownMenuBox
-import androidx.compose.material3.ExposedDropdownMenuDefaults
-import androidx.compose.material3.MenuAnchorType
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import de.fampopprol.dhbwhorb.R
-import de.fampopprol.dhbwhorb.data.dualis.models.StudyGrades
-import de.fampopprol.dhbwhorb.data.dualis.models.Module
-import de.fampopprol.dhbwhorb.data.dualis.models.ExamState
-import de.fampopprol.dhbwhorb.data.dualis.models.Semester
+import de.fampopprol.dhbwhorb.data.cache.GradesCacheManager
 import de.fampopprol.dhbwhorb.data.dualis.network.DualisService
 import de.fampopprol.dhbwhorb.data.security.CredentialManager
-import de.fampopprol.dhbwhorb.data.cache.GradesCacheManager
-import kotlinx.coroutines.launch
-import java.util.Locale
+import de.fampopprol.dhbwhorb.ui.screen.gradesScreen.GradesAuthManager
+import de.fampopprol.dhbwhorb.ui.screen.gradesScreen.GradesContent
+import de.fampopprol.dhbwhorb.ui.screen.gradesScreen.GradesDataManager
+import de.fampopprol.dhbwhorb.ui.screen.gradesScreen.GradesStateManager
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -72,199 +53,37 @@ fun GradesScreen(
     gradesCacheManager: GradesCacheManager? = null
 ) {
     val scope = rememberCoroutineScope()
-    var isLoading by remember { mutableStateOf(true) }
-    var studyGrades by remember { mutableStateOf<StudyGrades?>(null) }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
-    var isRefreshing by remember { mutableStateOf(false) }
     val pullRefreshState = rememberPullToRefreshState()
 
-    // Semester selection state
-    var availableSemesters by remember { mutableStateOf<List<Semester>>(emptyList()) }
-    var selectedSemester by remember { mutableStateOf<Semester?>(null) }
-    var isLoadingSemesters by remember { mutableStateOf(true) }
+    // Create managers
+    val dataManager = remember {
+        GradesDataManager(dualisService, gradesCacheManager, scope)
+    }
 
-    // Store string resources in variables that can be accessed from non-composable functions
-    val failedToLoadGradesMessage = stringResource(R.string.failed_to_load_grades)
+    val authManager = remember {
+        GradesAuthManager(dualisService, credentialManager, scope)
+    }
+
+    val stateManager = remember {
+        GradesStateManager(dataManager, authManager, scope)
+    }
+
+    // Error messages from string resources
     val authenticationFailedMessage = stringResource(R.string.authentication_failed)
     val noCredentialsFoundMessage = stringResource(R.string.no_credentials_found)
     val pleaseLoginMessage = stringResource(R.string.please_login)
+    val failedToLoadGradesMessage = stringResource(R.string.failed_to_load_grades)
 
-    // Function to fetch grades data for a specific semester
-    fun fetchGradesForSemester(
-        semester: Semester,
-        updateLoadingState: Boolean = true,
-        forceRefresh: Boolean = false
-    ) {
-        if (updateLoadingState && !isRefreshing) {
-            isLoading = true
-        }
-        errorMessage = null
-
-        scope.launch {
-            // Check cache first if not forcing refresh
-            if (!forceRefresh && gradesCacheManager != null) {
-                if (gradesCacheManager.hasValidCachedGrades(semester)) {
-                    val cachedData = gradesCacheManager.getCachedGrades()
-                    if (cachedData != null) {
-                        studyGrades = cachedData.first
-                        selectedSemester = cachedData.second
-                        isLoading = false
-                        isRefreshing = false
-                        Log.d(
-                            "GradesScreen",
-                            "Using cached grades for semester: ${semester.displayName}"
-                        )
-                        return@launch
-                    }
-                }
-            }
-
-            // Fetch from network
-            dualisService.getStudyGradesForSemester(semester) { result ->
-                isLoading = false
-                isRefreshing = false
-
-                if (result != null) {
-                    studyGrades = result
-                    selectedSemester = semester
-                    errorMessage = null
-
-                    // Cache the result
-                    scope.launch {
-                        gradesCacheManager?.cacheGrades(result, semester)
-                    }
-
-                    Log.d(
-                        "GradesScreen",
-                        "Fetched grades for semester ${semester.displayName}: $result"
-                    )
-                } else {
-                    errorMessage = failedToLoadGradesMessage
-                    Log.e(
-                        "GradesScreen",
-                        "Failed to fetch grades for semester ${semester.displayName}"
-                    )
-                }
-            }
-        }
-    }
-
-    // Function to fetch available semesters
-    fun fetchSemesters(forceRefresh: Boolean = false) {
-        isLoadingSemesters = true
-
-        scope.launch {
-            // Check cache first if not forcing refresh
-            if (!forceRefresh && gradesCacheManager != null) {
-                val cachedSemesters = gradesCacheManager.getCachedSemesters()
-                if (cachedSemesters != null) {
-                    availableSemesters = cachedSemesters
-                    isLoadingSemesters = false
-
-                    // Select the appropriate semester
-                    val defaultSemester =
-                        cachedSemesters.find { it.isSelected } ?: cachedSemesters.first()
-                    if (selectedSemester == null) {
-                        selectedSemester = defaultSemester
-                        fetchGradesForSemester(defaultSemester, false)
-                    }
-
-                    Log.d(
-                        "GradesScreen",
-                        "Using cached semesters: ${cachedSemesters.size} semesters"
-                    )
-                    return@launch
-                }
-            }
-
-            // Fetch from network
-            dualisService.getAvailableSemesters { semesters ->
-                isLoadingSemesters = false
-                if (semesters != null && semesters.isNotEmpty()) {
-                    availableSemesters = semesters
-
-                    // Cache the semesters
-                    scope.launch {
-                        gradesCacheManager?.cacheSemesters(semesters)
-                    }
-
-                    // Select the first semester that is marked as selected, or the first one if none are selected
-                    val defaultSemester = semesters.find { it.isSelected } ?: semesters.first()
-                    selectedSemester = defaultSemester
-                    // Fetch grades for the default semester
-                    fetchGradesForSemester(defaultSemester, false, forceRefresh)
-                    Log.d(
-                        "GradesScreen",
-                        "Fetched ${semesters.size} semesters, selected: ${defaultSemester.displayName}"
-                    )
-                } else {
-                    // Fallback to default semester selection if fetching fails
-                    val defaultSemesters = Semester.getDefaultSemesters()
-                    availableSemesters = defaultSemesters
-                    val defaultSemester =
-                        defaultSemesters.find { it.isSelected } ?: defaultSemesters.first()
-                    selectedSemester = defaultSemester
-                    fetchGradesForSemester(defaultSemester, false, forceRefresh)
-                    Log.w("GradesScreen", "Failed to fetch semesters, using defaults")
-                }
-            }
-        }
-    }
-
-    // Function to ensure authentication before fetching data
-    fun ensureAuthAndFetchData(forceRefresh: Boolean = false) {
-        if (!isRefreshing) {
-            isLoading = true
-        }
-        errorMessage = null
-
-        // Check if we need to authenticate first
-        if (!dualisService.isAuthenticated() && credentialManager != null) {
-            scope.launch {
-                val hasStoredCredentials = credentialManager.hasStoredCredentialsBlocking()
-                if (hasStoredCredentials) {
-                    val username = credentialManager.getUsernameBlocking()
-                    val password = credentialManager.getPassword()
-
-                    if (username != null && password != null) {
-                        Log.d("GradesScreen", "Re-authenticating before fetching data")
-                        dualisService.login(username, password) { result ->
-                            if (result != null) {
-                                Log.d(
-                                    "GradesScreen",
-                                    "Authentication successful, now fetching semesters"
-                                )
-                                // Now we can fetch semesters and grades after successful authentication
-                                fetchSemesters(forceRefresh)
-                            } else {
-                                isLoading = false
-                                isRefreshing = false
-                                errorMessage = authenticationFailedMessage
-                                Log.e("GradesScreen", "Authentication failed")
-                            }
-                        }
-                    } else {
-                        isLoading = false
-                        isRefreshing = false
-                        errorMessage = noCredentialsFoundMessage
-                        Log.e("GradesScreen", "No stored credentials found")
-                    }
-                } else {
-                    isLoading = false
-                    isRefreshing = false
-                    errorMessage = pleaseLoginMessage
-                    Log.e("GradesScreen", "No stored credentials found")
-                }
-            }
-        } else {
-            // Already authenticated, fetch semesters and grades directly
-            fetchSemesters(forceRefresh)
-        }
-    }
-
-    // Initial data loading
+    // Initialize on first composition
     LaunchedEffect(Unit) {
-        ensureAuthAndFetchData()
+        stateManager.initialize(
+            failedToLoadGradesMessage = failedToLoadGradesMessage,
+            authenticationFailedMessage = authenticationFailedMessage,
+            noCredentialsFoundMessage = noCredentialsFoundMessage,
+            pleaseLoginMessage = pleaseLoginMessage
+        ) { authResult ->
+            // Error handling is now done in the initialize method
+        }
     }
 
     Box(
@@ -275,12 +94,14 @@ fun GradesScreen(
         PullToRefreshBox(
             state = pullRefreshState,
             onRefresh = {
-                isRefreshing = true
-                selectedSemester?.let { semester ->
-                    fetchGradesForSemester(semester, false, forceRefresh = true)
-                } ?: ensureAuthAndFetchData(forceRefresh = true)
+                stateManager.onRefresh(
+                    failedToLoadGradesMessage = failedToLoadGradesMessage,
+                    authenticationFailedMessage = authenticationFailedMessage,
+                    noCredentialsFoundMessage = noCredentialsFoundMessage,
+                    pleaseLoginMessage = pleaseLoginMessage
+                )
             },
-            isRefreshing = isRefreshing,
+            isRefreshing = stateManager.isRefreshing,
             modifier = Modifier.fillMaxSize(),
         ) {
             Box(
@@ -288,532 +109,100 @@ fun GradesScreen(
                     .fillMaxSize()
                     .verticalScroll(rememberScrollState())
             ) {
-                if (isLoading && !isRefreshing) {
-                    // Show loading indicator when initially loading
-                    Box(
-                        modifier = Modifier.fillMaxSize(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            CircularProgressIndicator(
-                                modifier = Modifier.padding(16.dp)
-                            )
-                            Text(
-                                text = if (isLoadingSemesters) {
-                                    stringResource(R.string.loading_semesters)
-                                } else {
-                                    stringResource(R.string.loading_grades)
-                                },
-                                style = MaterialTheme.typography.bodyMedium
-                            )
-                        }
-                    }
-                } else if (errorMessage != null && studyGrades == null) {
-                    // Show error message
-                    Column(
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .padding(24.dp),
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                        verticalArrangement = Arrangement.Center
-                    ) {
-                        Text(
-                            text = errorMessage!!,
-                            color = MaterialTheme.colorScheme.error,
-                            style = MaterialTheme.typography.bodyLarge,
-                            textAlign = TextAlign.Center
+                when {
+                    stateManager.isLoading && !stateManager.isRefreshing -> {
+                        LoadingContent(
+                            isLoadingSemesters = stateManager.isLoadingSemesters
                         )
-                        Spacer(modifier = Modifier.height(16.dp))
-                        Button(
-                            onClick = { ensureAuthAndFetchData(forceRefresh = true) }
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.Refresh,
-                                contentDescription = stringResource(R.string.retry)
-                            )
-                            Spacer(modifier = Modifier.padding(horizontal = 4.dp))
-                            Text(stringResource(R.string.retry))
-                        }
                     }
-                } else {
-                    // Show grades content with semester selection
-                    GradesContent(
-                        studyGrades = studyGrades,
-                        isRefreshing = isRefreshing,
-                        availableSemesters = availableSemesters,
-                        selectedSemester = selectedSemester,
-                        onSemesterSelected = { semester ->
-                            selectedSemester = semester
-                            fetchGradesForSemester(semester)
-                        },
-                        modifier = Modifier.fillMaxSize()
-                    )
-                }
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun GradesContent(
-    studyGrades: StudyGrades?,
-    isRefreshing: Boolean,
-    availableSemesters: List<Semester>,
-    selectedSemester: Semester?,
-    onSemesterSelected: (Semester) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier.padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        // Show a thin progress indicator at top during refresh
-        if (isRefreshing) {
-            LinearProgressIndicator(
-                modifier = Modifier.fillMaxWidth(),
-                color = MaterialTheme.colorScheme.primary
-            )
-        } else {
-            // Empty spacer to maintain layout stability
-            Spacer(modifier = Modifier.height(4.dp))
-        }
+                    stateManager.errorMessage != null && stateManager.studyGrades == null -> {
+                        ErrorContent(
+                            errorMessage = stateManager.errorMessage!!,
+                            onRetry = {
+                                stateManager.onRetry(
+                                    failedToLoadGradesMessage = failedToLoadGradesMessage,
+                                    authenticationFailedMessage = authenticationFailedMessage,
+                                    noCredentialsFoundMessage = noCredentialsFoundMessage,
+                                    pleaseLoginMessage = pleaseLoginMessage
+                                )
+                            }
+                        )
+                    }
 
-        // Grades Header
-        Text(
-            text = stringResource(R.string.your_study_performance),
-            style = MaterialTheme.typography.headlineSmall,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(8.dp)
-        )
-
-        // Semester Selection
-        Column {
-            Text(
-                text = stringResource(R.string.select_semester),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                modifier = Modifier.padding(vertical = 8.dp)
-            )
-
-            // Semester dropdown
-            var expanded by remember { mutableStateOf(false) }
-
-            ExposedDropdownMenuBox(
-                expanded = expanded,
-                onExpandedChange = { expanded = !expanded }
-            ) {
-                OutlinedTextField(
-                    value = selectedSemester?.displayName ?: "",
-                    onValueChange = { },
-                    readOnly = true,
-                    label = { Text(stringResource(R.string.select_semester)) },
-                    trailingIcon = {
-                        ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded)
-                    },
-                    colors = ExposedDropdownMenuDefaults.outlinedTextFieldColors(),
-                    modifier = Modifier
-                        .menuAnchor(MenuAnchorType.PrimaryNotEditable)
-                        .fillMaxWidth()
-                )
-
-                ExposedDropdownMenu(
-                    expanded = expanded,
-                    onDismissRequest = { expanded = false }
-                ) {
-                    // Menu items for each available semester
-                    availableSemesters.forEach { semester ->
-                        DropdownMenuItem(
-                            onClick = {
-                                onSemesterSelected(semester)
-                                expanded = false
+                    else -> {
+                        GradesContent(
+                            studyGrades = stateManager.studyGrades,
+                            isRefreshing = stateManager.isRefreshing,
+                            availableSemesters = stateManager.availableSemesters,
+                            selectedSemester = stateManager.selectedSemester,
+                            onSemesterSelected = { semester ->
+                                stateManager.onSemesterSelected(
+                                    semester = semester,
+                                    failedToLoadGradesMessage = failedToLoadGradesMessage
+                                )
                             },
-                            text = {
-                                Text(
-                                    text = semester.displayName,
-                                    style = MaterialTheme.typography.bodyMedium
-                                )
-                            }
+                            modifier = Modifier.fillMaxSize()
                         )
                     }
                 }
             }
         }
-
-        // GPA Card with enhanced Material 3 styling
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.primaryContainer,
-                contentColor = MaterialTheme.colorScheme.onPrimaryContainer
-            ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 6.dp)
-        ) {
-            Column(
-                modifier = Modifier.padding(20.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Star,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.primary,
-                        modifier = Modifier.size(24.dp)
-                    )
-                    Text(
-                        text = stringResource(R.string.grade_point_average),
-                        style = MaterialTheme.typography.titleLarge,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-
-                Card(
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.surface,
-                        contentColor = MaterialTheme.colorScheme.onSurface
-                    ),
-                    elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-                ) {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = stringResource(R.string.overall_gpa),
-                            style = MaterialTheme.typography.bodyLarge,
-                            fontWeight = FontWeight.Medium
-                        )
-                        Text(
-                            text = studyGrades?.gpaTotal?.toString()
-                                ?: stringResource(R.string.not_available),
-                            style = MaterialTheme.typography.headlineSmall,
-                            fontWeight = FontWeight.Bold,
-                            color = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-            }
-        }
-
-        // Credits Card with enhanced styling and progress visualization
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.secondaryContainer,
-                contentColor = MaterialTheme.colorScheme.onSecondaryContainer
-            ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 6.dp)
-        ) {
-            Column(
-                modifier = Modifier.padding(20.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Star, // You might want to use a different icon
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.secondary,
-                        modifier = Modifier.size(24.dp)
-                    )
-                    Text(
-                        text = stringResource(R.string.credit_points),
-                        style = MaterialTheme.typography.titleLarge,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-
-                // Credits info cards
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Card(
-                        modifier = Modifier.weight(1f),
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surface,
-                            contentColor = MaterialTheme.colorScheme.onSurface
-                        ),
-                        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(16.dp),
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            Text(
-                                text = stringResource(R.string.credits_gained),
-                                style = MaterialTheme.typography.labelMedium,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                textAlign = TextAlign.Center,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                            Text(
-                                text = studyGrades?.creditsGained?.toString() ?: "0",
-                                style = MaterialTheme.typography.headlineSmall,
-                                fontWeight = FontWeight.Bold,
-                                color = MaterialTheme.colorScheme.primary,
-                                textAlign = TextAlign.Center,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                        }
-                    }
-
-                    Card(
-                        modifier = Modifier.weight(1f),
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surface,
-                            contentColor = MaterialTheme.colorScheme.onSurface
-                        ),
-                        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(16.dp),
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            Text(
-                                text = stringResource(R.string.credits_total),
-                                style = MaterialTheme.typography.labelMedium,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                textAlign = TextAlign.Center,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                            Text(
-                                text = studyGrades?.creditsTotal?.toString() ?: "0",
-                                style = MaterialTheme.typography.headlineSmall,
-                                fontWeight = FontWeight.Bold,
-                                color = MaterialTheme.colorScheme.secondary,
-                                textAlign = TextAlign.Center,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                        }
-                    }
-                }
-
-                // Enhanced Progress Bar with Material 3 styling
-                if (studyGrades != null) {
-                    val progress = if (studyGrades.creditsTotal > 0) {
-                        (studyGrades.creditsGained / studyGrades.creditsTotal).toFloat()
-                    } else 0f
-
-                    Card(
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surface,
-                            contentColor = MaterialTheme.colorScheme.onSurface
-                        ),
-                        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier.padding(16.dp)
-                        ) {
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceBetween,
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Text(
-                                    text = "Study Progress",
-                                    style = MaterialTheme.typography.labelLarge,
-                                    fontWeight = FontWeight.Medium
-                                )
-                                Text(
-                                    text = String.format(
-                                        Locale.getDefault(),
-                                        "%.1f%%",
-                                        progress * 100
-                                    ),
-                                    style = MaterialTheme.typography.labelLarge,
-                                    fontWeight = FontWeight.Bold,
-                                    color = MaterialTheme.colorScheme.primary
-                                )
-                            }
-
-                            Spacer(modifier = Modifier.height(8.dp))
-
-                            LinearProgressIndicator(
-                                progress = { progress },
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(8.dp),
-                                color = MaterialTheme.colorScheme.primary,
-                                trackColor = MaterialTheme.colorScheme.surfaceVariant,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-
-        // Individual Modules Section
-        if (studyGrades != null && studyGrades.modules.isNotEmpty()) {
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.surface,
-                    contentColor = MaterialTheme.colorScheme.onSurface
-                ),
-                elevation = CardDefaults.cardElevation(defaultElevation = 0.dp)
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.individual_modules),
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Bold
-                    )
-
-                    studyGrades.modules.forEach { module ->
-                        ModuleCard(module = module)
-                    }
-                }
-            }
-        }
-
-        // Additional info message
-        Card(
-            modifier = Modifier.fillMaxWidth(),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface,
-                contentColor = MaterialTheme.colorScheme.onSurface
-            ),
-            elevation = CardDefaults.cardElevation(defaultElevation = 0.dp)
-        ) {
-            Row(
-                modifier = Modifier.padding(16.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Star,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.padding(end = 12.dp)
-                )
-
-                Text(
-                    text = stringResource(R.string.detailed_module_info),
-                    style = MaterialTheme.typography.bodyMedium
-                )
-            }
-        }
-
-        Spacer(modifier = Modifier.height(32.dp))
     }
 }
 
 @Composable
-fun ModuleCard(
-    module: Module,
-    modifier: Modifier = Modifier
+private fun LoadingContent(
+    isLoadingSemesters: Boolean
 ) {
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = when (module.state) {
-                ExamState.PASSED -> MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-                ExamState.FAILED -> MaterialTheme.colorScheme.errorContainer.copy(alpha = 0.3f)
-                ExamState.PENDING -> MaterialTheme.colorScheme.surfaceVariant
-            },
-            contentColor = when (module.state) {
-                ExamState.PASSED -> MaterialTheme.colorScheme.onPrimaryContainer
-                ExamState.FAILED -> MaterialTheme.colorScheme.onErrorContainer
-                ExamState.PENDING -> MaterialTheme.colorScheme.onSurfaceVariant
-            }
-        ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp)
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
     ) {
         Column(
-            modifier = Modifier.padding(12.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.Top
-            ) {
-                Column(
-                    modifier = Modifier.weight(1f),
-                    verticalArrangement = Arrangement.spacedBy(4.dp)
-                ) {
-                    Text(
-                        text = module.name,
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.Medium
-                    )
-                    Text(
-                        text = module.id,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
+            CircularProgressIndicator(
+                modifier = Modifier.padding(16.dp)
+            )
+            Text(
+                text = if (isLoadingSemesters) {
+                    stringResource(R.string.loading_semesters)
+                } else {
+                    stringResource(R.string.loading_grades)
+                },
+                style = MaterialTheme.typography.bodyMedium
+            )
+        }
+    }
+}
 
-                Column(
-                    horizontalAlignment = Alignment.End,
-                    verticalArrangement = Arrangement.spacedBy(4.dp)
-                ) {
-                    Text(
-                        text = if (module.grade == "noch nicht gesetzt") stringResource(R.string.not_graded) else module.grade,
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.Bold,
-                        color = when (module.state) {
-                            ExamState.PASSED -> MaterialTheme.colorScheme.primary
-                            ExamState.FAILED -> MaterialTheme.colorScheme.error
-                            ExamState.PENDING -> MaterialTheme.colorScheme.onSurfaceVariant
-                        }
-                    )
-                    Text(
-                        text = "${module.credits} ${stringResource(R.string.credit_points_short)}",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            }
-
-            // Status indicator
-            val statusText = when (module.state) {
-                ExamState.PASSED -> stringResource(R.string.passed)
-                ExamState.FAILED -> stringResource(R.string.failed)
-                ExamState.PENDING -> stringResource(R.string.pending)
-            }
-
-            val statusIcon = when (module.state) {
-                ExamState.PASSED -> "✓"
-                ExamState.FAILED -> "✗"
-                ExamState.PENDING -> "⏱"
-            }
-
-            Row(
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = statusIcon,
-                    style = MaterialTheme.typography.bodySmall,
-                    modifier = Modifier.padding(end = 4.dp)
-                )
-                Text(
-                    text = statusText,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = when (module.state) {
-                        ExamState.PASSED -> MaterialTheme.colorScheme.primary
-                        ExamState.FAILED -> MaterialTheme.colorScheme.error
-                        ExamState.PENDING -> MaterialTheme.colorScheme.onSurfaceVariant
-                    }
-                )
-            }
+@Composable
+private fun ErrorContent(
+    errorMessage: String,
+    onRetry: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(24.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Text(
+            text = errorMessage,
+            color = MaterialTheme.colorScheme.error,
+            style = MaterialTheme.typography.bodyLarge,
+            textAlign = TextAlign.Center
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        Button(onClick = onRetry) {
+            Icon(
+                imageVector = Icons.Default.Refresh,
+                contentDescription = stringResource(R.string.retry)
+            )
+            Spacer(modifier = Modifier.padding(horizontal = 4.dp))
+            Text(stringResource(R.string.retry))
         }
     }
 }
